fix python for structured var
There was a problem with the asumptions when building the original
Pythons script: that all variables needed to be enclosed in quotes in
the terraform.tfvars file. It turns out that sometimes, we have
variables that shouldn't be quoted, in particular, variables that are a
collection of data inside of curly braces.

This version of the aws_parse.py script includes and if/then/else
to test for an opening left curly brace ({) in the variable and if it
exists, it writes the value without quotes.

The terraform code for apps/app2 was updated to read in variables
with this kind of data and write the result as an output. Terraform plan
will not show any changes, but a successful terraform plan (with no
changes) will indicate that it read in the variables correctly.