Update aws_parse.py
Discovered that there are other leading characters in variables that should not be quoted in the .tfvars output. 

The logic was updated from checking if the first character was a curly left brace to checking if the first character is either a curly left brace or a left square bracket.