Automate network creation and destruction
There are costs for running our full network infrastructure in AWS: the
Internet Gateway, NAT Gateway, and Elastic IPs are chargeable
resources. These expenses leave us less room for testing in our
restricted budget for this environment.

Leverage the GitHub Actions Scheduler.  Two new workflows (for
`terraform apply` and `terraform destroy` are included. Each is
configured to run on a schedul (apply at 7am, destroy at 7pm). This
means that the network infrastructure charges should be less
than 50% of the previous costs for running 24x7.

This does have the side effect of recreating new elastic IPs
for the "static IP" resources each day. With this only being a test/dev
environment, this shouldn't be a problem. We can also
disable these workflows if we need to keep the network infrastructure
running for longer than one day.