Fix branch name problem
The DETACHED HEAD git checkout breaks the Python
git integration in regard to branch name (there is
no branch name for a detached head situation).

GHA has a solution to this problem since it captures the
branch name in the GITHUB_REF GHA environment variable. This
version of the code is an attempt to leverage that information to
fix the action.