Test all variables for all subfolders
Needed to verify both DEV and GLOBAL `terraform plan`
actions.

This triggers updates in each subfolder for a complete test of
the GHA `terraform plan` actions. This commit was preceeded by running
the `parameter_update.py` script in each subfolder for both the
active branch and for the `main` branch for all the
various Terraform workspaces.