Update repo for new Python scripts for improved GHA workflows
The following changes were made to the handling of variables for improved GitHub Actions automation.

* Update scripts for variables

There were issues with the hierarchy of variables in Parameter Store, especially related to distinguishing variables between repos and distinguishing variables between branches in a single repo.

Eric built updated scripts in Python to handle the new hierarchy. The corresponding GitHub Actions workflows needed to be updated to use the new Parameter Store paths. This updates both tf-validate and tf-plan-dev to use the updated paths. It also cleans up the GitHub Action to run both the Python and the Terraform without having to move around files after pulling down the repo from GitHub.

* Fix pipenv run python syntax

The Python code takes care of constructing the full path to the variables in Parameter Store. I only need to pass the
<folder><subfolder> path to the script.

* Fix branch name problem

The DETACHED HEAD git checkout breaks the Python git integration in regard to branch name (there is
no branch name for a detached head situation).

The GITHUB_HEAD_REF default environment variable correctly captures the name of the branch connected to the PR.

* Clean up documentation and workflows

The branch name problem is solved, but the workflow files need some cleaning up.

Updates to the `./scripts/` README file to capture the new branch_name command line parameter for the Python scripts. Updates to both the DEV and GLOBAL `terraform plan` workflows to use the updated scripts correctly.

* Fix the apply workflows

The `terraform apply` GHA workflows needed updating to work with the new Python scripts.

The GHA workflows for `terraform apply` are updated to use the new scripts and some typos in the workflows were addressed.