Update GutHub Actions to Terraform v0.13
Why these changes are being introduced:
InfraEng is updating all our Terraform code to be compatible with
Terraform version 0.13. This requires that all of our GitHub Actions
automations need to run a 0.13.x version of Terraform, currently
v0.13.7.

How this addresses that need:
* Updates the main README and the scripts/README to reflect the
new version of Terraform used by the GitHub Actions
* Updates each .yml file in .github/workflows to install version 0.13.7
of the Terraform command
* Updates the tf-network-apply.yml and tf-network-destroy.yml workflows
to use the most recent versions of the Python variable management
scripts (this had not been done previously)
* Removes the deprecated aws_parse.py script from scripts/

Side effects of this change:
* For anyone running this code locally, prior to using Terraform 0.13.7,
the .terraform subdirectory should be removed.
* Terraform 0.13.7 handles provider versions differently, so any
`terraform plan` and `terraform apply` runs with the current version
of our code might cause a newer version of the AWS Provider to be
used.

Since the code in this repository does nothing other than produce
outputs, there will be no issues with any version incompatibilities
of our Terraform code and running `terraform plan` and `terraform apply`.
When we introduce these same changes in other repositories, we will
need to be MUCH more careful!

https://mitlibraries.atlassian.net/browse/IN-240

On branch gha-tf13-upgrade
Changes to be committed:
	modified:   .github/workflows/tf-apply-dev.yml
	modified:   .github/workflows/tf-apply-global.yml
	modified:   .github/workflows/tf-network-apply.yml
	modified:   .github/workflows/tf-network-destroy.yml
	modified:   .github/workflows/tf-plan-dev.yml
	modified:   .github/workflows/tf-plan-global.yml
	modified:   .github/workflows/tf-promote.yml
	modified:   .github/workflows/tf-validate.yml
	modified:   README.md
	modified:   scripts/README.md
	deleted:    scripts/aws_parse.py