GitHub Action - fix tf-validate on initial push (#11)
* Fix terraform validate for initial push

Why these changes are being introduced:
The tf-validate workflow (since it was first introduced to replace the
TravisCI automation) has always failed on the very first push of a new
branch to GitHub. This update addresses that problem.

How this addresses that need:
* Remove the `steps.if: ...` logic in the "generate matrix" job
* Add `if .. then .. else .. fi` logic to the bash commands in the
"generate matrix" job's `run:` command.
* On first push (when $github.event.before=all-zeros), compare the new
branch to `origin/main` with the `git diff` command
* Update the `validate` jobs to reflect the changes that were
implemented in the terraform-testdev-* and mitlib-terraform
repositories.

Additionally, the in-line documentation was updated to reflect the same
documentation that is in the Dev/Test repos and the Stage/Prod repo.

Side effects of this change:
There are no side effects of this change. It does assume that all new
branches created by developers are branch from `origin/main`.