Major clean of GHA workflows
Why these changes are being introduced:
In the process of testing and updating workflows for the update to
Terraform 0.13, some inconsistencies were introduced in the various
GitHub Actions workflows in this testing repository. Additionally, some
gaps in the documentation were found.

How this addresses that need:
* a README file now exists in the `.github/workflows` folder that has
details on how to do local testing of some of the workflow commands
* all the in-line documentation in all of the workflows has been
updated and improved
* the filtering (awk/sed/grep) of the `git diff` has been cleaned up
and is now consistent across all of the workflows
* the workflow triggers (on.push, on.pull_request, on.workflow_dispatch)
have all been cleaned up for consistency
* the workflows have all been updated to only run Terraform 0.13.7

Side effects of this change:
None.

Changes to be committed:
	new file:   .github/workflows/README.md
	modified:   .github/workflows/tf-apply-dev.yml
	modified:   .github/workflows/tf-apply-global.yml
	modified:   .github/workflows/tf-plan-dev.yml
	modified:   .github/workflows/tf-plan-global.yml
	modified:   .github/workflows/tf-promote.yml
	modified:   .github/workflows/tf-validate.yml