Upgrade for Terraform 0.13 (#12)
Why these changes are being introduced:
We are updating the code in all of our Terraform repositories for
compatibility with Terraform 0.13. Tweaks to the GHA workflows 
(tf-validate, tf-plan-dev, tf-plan-global) were made when rolling 
out the updates to the other Terraform repositories 
(terraform-testdev-core, terraform-testdev-apps, and
mitlib-terraform). These "downstream" changes needed to be 
rolled into our GHA testing repository for consistency across 
repos. In the process of testing and updating workflows for the update 
to Terraform 0.13, some inconsistencies were introduced in the various
GitHub Actions workflows in this testing repository. Additionally, some
gaps in the documentation were found.

* Upgrade shared/network to 0.13
* Upgrade shared/core to 0.13
* Upgrade modules/tf-mod-shared-core to 0.13
* Upgrade modules/tf-mod-shared-network to 0.13
* Upgrade apps/app1 to 0.13
* Upgrade apps/app2 to 0.13
* Update GHA workflows and improve documentation


How this addresses that need:
* Updates the module call for tf-mod-shared to tag=0.13
* Adds the `required_providers` block to the `terraform { }` block
in `versions.tf`
* Adds type constraints to variable declarations (as needed)
* Fix required_version constraint in the terraform { } block (as needed)
* update module call for tf-mod-name for 0.13 tag
* tf-validate: Include the TF_WORKSPACE environement variable
* tf-validate: Update in-line documentation
* tf-plan-dev: Update `on.pull-request` from `paths-ignore` filtering to
`paths` filtering
* tf-plan-global: Update `on.pull-request` from `paths-ignore`
filtering to `paths` filtering
* the tf-promote workflow was change to only allow for a manual
trigger
* the tf-apply-dev workflow was changed for on.pull-request.paths
instead of on.pull-request.paths-ignore
* a README file now exists in the `.github/workflows` folder that has
details on how to do local testing of some of the workflow commands
* all the in-line documentation in all of the workflows has been
updated and improved
* the filtering (awk/sed/grep) of the `git diff` has been cleaned up
and is now consistent across all of the workflows
* the workflow triggers (on.push, on.pull_request, on.workflow_dispatch)
have all been cleaned up for consistency
* the workflows have all been updated to only run Terraform 0.13.7


Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-240